Restrukturierung für Verwendung der Filmsuche nach Kategorien
