+++ src/main/java/mServer/crawler/sender/arte/ArteVideoDetailsDTO.java
@@ -5,7 +5,12 @@ import mServer.crawler.sender.newsearch.GeoLocations;
 public class ArteVideoDetailsDTO {
     private String broadcastBegin = "";
     private GeoLocations geo = GeoLocations.GEO_NONE;
-    
+
+    private String title = "";
+    private String theme = "";
+    private String description = "";
+    private String website = "";
+
     public String getBroadcastBegin() {
         return this.broadcastBegin;
     }
@@ -20,5 +25,30 @@ public class ArteVideoDetailsDTO {
     
     public void setGeoLocation(GeoLocations aGeo) {
         this.geo = aGeo;
+    }
+    
+    public String getTitle() {
+        return title;
+    }
+    public void setTitle(String aTitle) {
+        title = aTitle;
+    }
+    public void setTheme(String aTheme) {
+        theme = aTheme;
+    }
+    public void setDescription(String aDescription) {
+        description = aDescription;
+    }
+    public void setWebsite(String aWebsite) {
+        website = aWebsite;
+    }
+    public String getTheme() {
+        return theme;
+    }
+    public String getDescription() {
+        return description;
+    }
+    public String getWebsite() {
+        return website;
     }
 }
